+++ CHANGES.md
@@ -8,6 +8,10 @@
 
 * Allow to construct `Data` instances to avoid double `Arc` for `Send + Sync` types.
 
+### Fixed
+
+* Fix `NormalizePath` middleware impl #806
+
 
 ## [1.0.0-beta.2] - 2019-04-24